refactor(client): 抽取自由窗口拖出检测与浮窗层到 sidebar/free-windows.tsx
Sidebar.tsx 拆分第五块：useFloatDragout（document 捕获态 dragover/
drop/dragend/blur 监听、会话列矩形命中测试、hint 状态去抖）与
FreeWindowLayer（floats.map 渲染 + 拖出 hint 覆盖层，fragment 渲染
保证 DOM 顺序与原先逐字一致）原样迁出，注释逐字保留。

行为不变：effect 依赖与回调闭包相同；层组件接收的 renderTab/
getTabIcon/store/ctx/sessionId/cwd 即原先 JSX 内联闭包捕获的同一批
每 render 新建的值，FreeWindow 收到的 props 与回调语义不变。
